// src/app/shared/models/flight.model.ts
export class Flight {
    idFlight: number;
    companyName: string | null;
    flightType: string | null;
    travelType: string | null;
    departureDate: Date | null;
    departureTime: Date | null;
    arrivalDate: Date | null;
    arrivalTime: Date | null;
    backDate: Date | null;
    backTime: Date | null;
    departureLocation: string | null;
    arrivalLocation: string | null;
    flightDuration: Date | null;
    connectionDuration: Date | null;
    aircraftType: string | null;
}